pyproject.toml
[project]
name = "steam-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "steamio",
]

[tool.uv.sources]
steamio = { git = "https://github.com/Gobot1234/steam.py.git" }
